colexec: put aggregate funcs in a separate file
This is just code movement and some files' renaming.

Release note: None
